Remove 'new file' update statistic
The update list implementation will be refactored to remove dependence
on the FROM file list and will no longer be able to differentiate between
new and changed files for update statistics reporting. The current update
list implementation will be kept as a fallback (when no minversion included
in the MoM header), so the 'new file' update statistic will be removed from
the current implementation to keep the statistic output consistent.

Also, this commit creates a separate header file (stats.h) for stats.c,
adds an enum type to manage swupd_stats indexes, and removes unused
statstics functions.
