+++ src/update.c
@@ -34,6 +34,7 @@
 
 #include "config.h"
 #include "signature.h"
+#include "stats.h"
 #include "swupd.h"
 
 static int requested_version = -1;
