+++ test/functional/update/update-boot-skip.bats
@@ -24,8 +24,7 @@ test_setup() {
 		    changed bundles   : 1
 		    new bundles       : 0
 		    deleted bundles   : 0
-		    changed files     : 0
-		    new files         : 4
+		    changed files     : 4
 		    deleted files     : 0
 		Starting download of remaining update content. This may take a while...
 		Finishing download of update content...
