+++ test/functional/update/update-include-old-bundle-with-tracked-file.bats
@@ -32,8 +32,7 @@ test_setup() {
 		    changed bundles   : 3
 		    new bundles       : 0
 		    deleted bundles   : 0
-		    changed files     : 2
-		    new files         : 4
+		    changed files     : 6
 		    deleted files     : 0
 		Starting download of remaining update content. This may take a while...
 		Finishing download of update content...
