+++ test/functional/update/update-iterative-bad-name.bats
@@ -58,7 +58,6 @@ test_setup() {
 		    new bundles       : 0
 		    deleted bundles   : 0
 		    changed files     : 3
-		    new files         : 0
 		    deleted files     : 0
 		Starting download of remaining update content. This may take a while...
 		Finishing download of update content...
