+++ test/functional/update/update-iterative-has-dependency.bats
@@ -35,7 +35,6 @@ test_setup() {
 		    new bundles       : 0
 		    deleted bundles   : 0
 		    changed files     : 2
-		    new files         : 0
 		    deleted files     : 0
 		Starting download of remaining update content. This may take a while...
 		Finishing download of update content...
