+++ test/functional/update/update-missing-os-core.bats
@@ -28,8 +28,7 @@ test_setup() {
 		    changed bundles   : 2
 		    new bundles       : 0
 		    deleted bundles   : 0
-		    changed files     : 1
-		    new files         : 1
+		    changed files     : 2
 		    deleted files     : 0
 		Starting download of remaining update content. This may take a while...
 		Finishing download of update content...
