+++ test/functional/update/update-newest-deleted.bats
@@ -31,7 +31,6 @@ test_setup() {
 		    new bundles       : 0
 		    deleted bundles   : 0
 		    changed files     : 1
-		    new files         : 0
 		    deleted files     : 1
 		Starting download of remaining update content. This may take a while...
 		Finishing download of update content...
