+++ test/functional/update/update-rename.bats
@@ -28,8 +28,7 @@ test_setup() {
 		    changed bundles   : 1
 		    new bundles       : 0
 		    deleted bundles   : 0
-		    changed files     : 0
-		    new files         : 1
+		    changed files     : 1
 		    deleted files     : 1
 		Starting download of remaining update content. This may take a while...
 		Finishing download of update content...
