+++ test/functional/update/update-skip-scripts.bats
@@ -26,8 +26,7 @@ test_setup() {
 		    changed bundles   : 1
 		    new bundles       : 0
 		    deleted bundles   : 0
-		    changed files     : 1
-		    new files         : 1
+		    changed files     : 2
 		    deleted files     : 0
 		Starting download of remaining update content. This may take a while...
 		Finishing download of update content...
