+++ test/functional/update/update-skip-verified-fullfiles.bats
@@ -34,7 +34,6 @@ test_setup() {
 		    new bundles       : 0
 		    deleted bundles   : 0
 		    changed files     : 1
-		    new files         : 0
 		    deleted files     : 0
 		Staging file content
 		Applying update
