+++ test/functional/update/update-with-old-mirror.bats
@@ -41,8 +41,7 @@ test_setup() {
 		    changed bundles   : 1
 		    new bundles       : 0
 		    deleted bundles   : 0
-		    changed files     : 0
-		    new files         : 2
+		    changed files     : 2
 		    deleted files     : 0
 		Starting download of remaining update content. This may take a while...
 		Finishing download of update content...
